+++ src/MeshBatcher.ts
@@ -11,15 +11,15 @@ import {
 } from 'oasis-engine';
 import { SkeletonMaterial } from './LottieMaterial';
 
+/**
+ * @internal
+ */
 export class MeshBatcher extends MeshRenderer {
-  private static VERTEX_SIZE = 9;
-
   vertices: Float32Array;
   indices: Uint16Array;
-
-  private indexBuffer: Buffer;
   vertexBuffer: Buffer;
 
+  private indexBuffer: Buffer;
   private verticesLength = 0;
   private indicesLength = 0;
 
@@ -27,7 +27,7 @@ export class MeshBatcher extends MeshRenderer {
     super(entity);
   }
 
-  init(l: number) {
+  init(l: number): void {
     const mesh = new BufferMesh(this.engine, 'lottie_batch_mesh');
 
     this.vertices = new Float32Array(36 * l);
@@ -69,9 +69,9 @@ export class MeshBatcher extends MeshRenderer {
     this.setMaterial(new SkeletonMaterial(this.engine));
   }
 
-  setBufferData() {
+  setBufferData(): void {
     const { indicesLength } = this;
-    let indexStart = this.verticesLength / MeshBatcher.VERTEX_SIZE;
+    let indexStart = this.verticesLength / 9;
 
     this.vertexBuffer.setData(this.vertices);
     this.verticesLength = this.verticesLength;
