+++ src/core/layer/SpriteLottieLayer.ts
@@ -1,13 +1,15 @@
 import { initialDefaultFrame } from '../contant';
 import BaseLottieLayer from './BaseLottieLayer';
 
+/**
+ * @internal
+ */
 export default class SpriteLottieLayer extends BaseLottieLayer {
-  private x: number;
-  private y: number;
+  x: number;
+  y: number;
 
   constructor(layer, atlas) {
     super(layer);
-
 
     if (layer.refId) {
       const { w, h, x, y } = atlas.frames[layer.refId + '.png'].frame;
